requirements.txt
anthropic>=0.25.0
typer>=0.12.0
rich>=13.0.0
httpx>=0.27.0
jinja2>=3.1.0
pyyaml>=6.0
